agent/pyproject.toml
[project]
name = "watchdog-agent"
version = "0.1.0"
description = "AI agent that monitors topics and notifies you of updates"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "dbus-python>=1.3.2",
    "schedule>=1.2.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.0.0",
]

[project.scripts]
watchdog-agent = "watchdog_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/watchdog_agent"]
